clean up ace.system.database ns
